Compile fix.
git-svn-id: file:///home/nick/svnrepo/repo/indigo/trunk@13794 20aa11e4-6e6f-594c-a032-6291ace9ee65

+++ graphics/KTXDecoder.cpp
@@ -103,7 +103,7 @@ Reference<Map2D> KTXDecoder::decode(const std::string& path)
 			//}
 
 			// Read mipPadding
-			file.setReadIndex(Maths::roundUpToMultipleOfPowerOf2(file.getReadIndex(), 4ull));
+			file.setReadIndex(Maths::roundUpToMultipleOfPowerOf2(file.getReadIndex(), (size_t)4));
 		}
 
 		return image;
